+++ core/src/test/java/com/google/adk/tools/retrieval/CassandraRagRetrievalTest.java
@@ -3,7 +3,6 @@
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * You may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
@@ -66,7 +65,7 @@ public final class CassandraRagRetrievalTest {
 
     // Arrange
 
-    List<Float> embedding = ImmutableList.of(0.1f, 0.2f, 0.3f);
+    List<Double> embedding = ImmutableList.of(0.1, 0.2, 0.3);
 
     Map<String, Object> args =
         ImmutableMap.of(
